Implement stable reuse hashes by sorting dictionary keys
Fixes #1553
